pyproject.toml
[project]
name = "glamira-ops-agent"
version = "0.1.0"
description = "GLAMIRA Ops Agent — autonomous AI operations coordinator"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43",
    "voyageai>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "redis[hiredis]>=5.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "google-auth>=2.37",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.159",
    "langfuse>=2.0",
    "structlog>=24.4",
    "python-dateutil>=2.9",
    "python-Levenshtein>=0.26",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent1"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
